refactor(test files): refactor test files [FINISHES #31]
